requirements.txt
Django==1.9.2
Markdown==2.6.5
MySQL-python==1.2.5
PyJWT==1.4.0
argparse==1.2.1
django-filters==0.2.1
django-rest-auth==0.6.0
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
six==1.10.0
wsgiref==0.1.2
